+++ internal/graph/check_test.go
@@ -148,7 +148,7 @@ func TestExclusionCheckFuncReducer(t *testing.T) {
 
 	ctx := context.Background()
 
-	concurrencyLimit := uint32(10)
+	concurrencyLimit := 10
 
 	t.Run("requires_exactly_two_handlers", func(t *testing.T) {
 		_, err := exclusion(ctx, concurrencyLimit)
@@ -478,6 +478,28 @@ func TestExclusionCheckFuncReducer(t *testing.T) {
 
 		wg.Wait() // just to make sure to avoid test leaks
 	})
+
+	t.Run("should_error_if_base_handler_panics", func(t *testing.T) {
+		panicHandler := func(context.Context) (*ResolveCheckResponse, error) {
+			panic(panicErr)
+		}
+
+		resp, err := exclusion(ctx, concurrencyLimit, panicHandler, falseHandler)
+		require.ErrorContains(t, err, panicErr)
+		require.ErrorIs(t, err, ErrPanic)
+		require.Nil(t, resp)
+	})
+
+	t.Run("should_error_if_sub_handler_panics", func(t *testing.T) {
+		panicHandler := func(context.Context) (*ResolveCheckResponse, error) {
+			panic(panicErr)
+		}
+
+		resp, err := exclusion(ctx, concurrencyLimit, trueHandler, panicHandler)
+		require.ErrorContains(t, err, panicErr)
+		require.ErrorIs(t, err, ErrPanic)
+		require.Nil(t, resp)
+	})
 }
 
 func TestIntersectionCheckFuncReducer(t *testing.T) {
@@ -487,7 +509,7 @@ func TestIntersectionCheckFuncReducer(t *testing.T) {
 
 	ctx := context.Background()
 
-	concurrencyLimit := uint32(10)
+	concurrencyLimit := 10
 
 	t.Run("no_handlers_return_false", func(t *testing.T) {
 		resp, err := intersection(ctx, concurrencyLimit)
@@ -1371,7 +1393,7 @@ func TestUnionCheckFuncReducer(t *testing.T) {
 
 	ctx := context.Background()
 
-	concurrencyLimit := uint32(10)
+	concurrencyLimit := 10
 
 	falseHandler := func(context.Context) (*ResolveCheckResponse, error) {
 		return &ResolveCheckResponse{
@@ -3251,7 +3273,7 @@ func TestProcessDispatch(t *testing.T) {
 				dispatchMsgChan <- dispatchMsg
 			}
 
-			outcomeChan := checker.processDispatches(ctx, uint32(tt.poolSize), dispatchMsgChan)
+			outcomeChan := checker.processDispatches(ctx, tt.poolSize, dispatchMsgChan)
 
 			// now, close the channel to simulate everything is sent
 			close(dispatchMsgChan)
@@ -3273,7 +3295,7 @@ func TestProcessDispatch(t *testing.T) {
 		checker.SetDelegate(mockResolver)
 
 		dispatchChan := make(chan dispatchMsg, 1)
-		outcomeChan := checker.processDispatches(ctx, uint32(1), dispatchChan)
+		outcomeChan := checker.processDispatches(ctx, 1, dispatchChan)
 		dispatchChan <- dispatchMsg{
 			dispatchParams: &dispatchParams{
 				parentReq: nil, // This will cause a panic when accessed in `dispatch`
@@ -3298,7 +3320,7 @@ func TestConsumeDispatch(t *testing.T) {
 	}
 	tests := []struct {
 		name                   string
-		limit                  uint32
+		limit                  int
 		ctxCancelled           bool
 		dispatchMsgs           []dispatchMsg
 		mockedDispatchResponse []*ResolveCheckResponse