+++ pkg/storage/storagewrappers/cached_datastore.go
@@ -3,13 +3,10 @@ package storagewrappers
 import (
 	"context"
 	"errors"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
 
-	"github.com/cespare/xxhash/v2"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"go.opentelemetry.io/otel"
@@ -35,19 +32,19 @@ var (
 		Namespace: build.ProjectName,
 		Name:      "tuples_cache_total_count",
 		Help:      "The total number of created cached iterator instances.",
-	}, []string{"operation"})
+	}, []string{"operation", "method"})
 
 	tuplesCacheHitCounter = promauto.NewCounterVec(prometheus.CounterOpts{
 		Namespace: build.ProjectName,
 		Name:      "tuples_cache_hit_count",
 		Help:      "The total number of cache hits from cached iterator instances.",
-	}, []string{"operation"})
+	}, []string{"operation", "method"})
 
 	tuplesCacheDiscardCounter = promauto.NewCounterVec(prometheus.CounterOpts{
 		Namespace: build.ProjectName,
 		Name:      "tuples_cache_discard_count",
 		Help:      "The total number of discards from cached iterator instances.",
-	}, []string{"operation"})
+	}, []string{"operation", "method"})
 
 	tuplesCacheSizeHistogram = promauto.NewHistogramVec(prometheus.HistogramOpts{
 		Namespace:                       build.ProjectName,
@@ -57,7 +54,7 @@ var (
 		NativeHistogramBucketFactor:     1.1,
 		NativeHistogramMaxBucketNumber:  100,
 		NativeHistogramMinResetDuration: time.Hour,
-	}, []string{"operation"})
+	}, []string{"operation", "method"})
 )
 
 // iterFunc is a function closure that returns an iterator
@@ -70,6 +67,13 @@ type CachedDatastoreOpt func(*CachedDatastore)
 func WithCachedDatastoreLogger(logger logger.Logger) CachedDatastoreOpt {
 	return func(b *CachedDatastore) {
 		b.logger = logger
+	}
+}
+
+// WithCachedDatastoreMethodName is used in metric differentiation to tell us if this was Check or ListObjects.
+func WithCachedDatastoreMethodName(method string) CachedDatastoreOpt {
+	return func(b *CachedDatastore) {
+		b.method = method
 	}
 }
 
@@ -91,6 +95,8 @@ type CachedDatastore struct {
 	wg *sync.WaitGroup
 
 	logger logger.Logger
+
+	method string // Whether this datastore is for Check or ListObjects
 }
 
 // NewCachedDatastore returns a wrapper over a datastore that caches iterators in memory.
@@ -113,6 +119,7 @@ func NewCachedDatastore(
 		sf:                      sf,
 		wg:                      wg,
 		logger:                  logger.NewNoopLogger(),
+		method:                  "",
 	}
 
 	for _, opt := range opts {
@@ -143,32 +150,23 @@ func (c *CachedDatastore) ReadStartingWithUser(
 		return iter(ctx)
 	}
 
-	var b strings.Builder
-	b.WriteString(
-		storage.GetReadStartingWithUserCacheKeyPrefix(store, filter.ObjectType, filter.Relation),
-	)
+	cacheKey, err := readStartingWithUserKey(store, filter)
+	if err != nil {
+		return nil, err
+	}
 
 	// NOTE: There is no need to limit the length of this
 	// since at most it will have 2 entries (user and wildcard if possible)
 	subjects := make([]string, 0, len(filter.UserFilter))
 	for _, objectRel := range filter.UserFilter {
-		subject := tuple.ToObjectRelationString(objectRel.GetObject(), objectRel.GetRelation())
-		subjects = append(subjects, subject)
-		b.WriteString("/" + subject)
-	}
-
-	if filter.ObjectIDs != nil {
-		hasher := xxhash.New()
-		for _, oid := range filter.ObjectIDs.Values() {
-			if _, err := hasher.WriteString(oid); err != nil {
-				return nil, err
-			}
+		subject := objectRel.GetObject()
+		if objectRel.GetRelation() != "" {
+			subject = tuple.ToObjectRelationString(objectRel.GetObject(), objectRel.GetRelation())
 		}
-
-		b.WriteString("/" + strconv.FormatUint(hasher.Sum64(), 10))
+		subjects = append(subjects, subject)
 	}
 
-	return c.newCachedIteratorByUserObjectType(ctx, "ReadStartingWithUser", store, iter, b.String(), subjects, filter.ObjectType)
+	return c.newCachedIteratorByUserObjectType(ctx, "ReadStartingWithUser", store, iter, cacheKey, subjects, filter.ObjectType)
 }
 
 // ReadUsersetTuples see [storage.RelationshipTupleReader].ReadUsersetTuples.
@@ -193,33 +191,13 @@ func (c *CachedDatastore) ReadUsersetTuples(
 		return iter(ctx)
 	}
 
-	var b strings.Builder
-	b.WriteString(
-		storage.GetReadUsersetTuplesCacheKeyPrefix(store, filter.Object, filter.Relation),
-	)
-
-	var rb strings.Builder
-	var wb strings.Builder
-
-	for _, userset := range filter.AllowedUserTypeRestrictions {
-		if _, ok := userset.GetRelationOrWildcard().(*openfgav1.RelationReference_Relation); ok {
-			rb.WriteString("/" + userset.GetType() + "#" + userset.GetRelation())
-		}
-		if _, ok := userset.GetRelationOrWildcard().(*openfgav1.RelationReference_Wildcard); ok {
-			wb.WriteString("/" + userset.GetType() + ":*")
-		}
-	}
-
-	// wildcard should have precedence
-	if wb.Len() > 0 {
-		b.WriteString(wb.String())
-	}
-
-	if rb.Len() > 0 {
-		b.WriteString(rb.String())
-	}
-
-	return c.newCachedIteratorByObjectRelation(ctx, "ReadUsersetTuples", store, iter, b.String(), filter.Object, filter.Relation)
+	return c.newCachedIteratorByObjectRelation(ctx,
+		"ReadUsersetTuples",
+		store,
+		iter,
+		readUsersetTuplesKey(store, filter),
+		filter.Object,
+		filter.Relation)
 }
 
 // Read see [storage.RelationshipTupleReader].Read.
@@ -249,11 +227,13 @@ func (c *CachedDatastore) Read(
 		return iter(ctx)
 	}
 
-	var b strings.Builder
-	b.WriteString(
-		storage.GetReadCacheKey(store, tuple.TupleKeyToString(tupleKey)),
-	)
-	return c.newCachedIteratorByObjectRelation(ctx, "Read", store, iter, b.String(), tupleKey.GetObject(), tupleKey.GetRelation())
+	return c.newCachedIteratorByObjectRelation(ctx,
+		"Read",
+		store,
+		iter,
+		readKey(store, tupleKey),
+		tupleKey.GetObject(),
+		tupleKey.GetRelation())
 }
 
 // findInCache tries to find a key in the cache.
@@ -374,10 +354,10 @@ func (c *CachedDatastore) newCachedIterator(
 ) (storage.TupleIterator, error) {
 	span := trace.SpanFromContext(ctx)
 	span.SetAttributes(attribute.String("cache_key", cacheKey))
-	tuplesCacheTotalCounter.WithLabelValues(operation).Inc()
+	tuplesCacheTotalCounter.WithLabelValues(operation, c.method).Inc()
 
 	if cacheEntry, ok := findInCache(c.cache, store, cacheKey, invalidEntityKeys, c.logger); ok {
-		tuplesCacheHitCounter.WithLabelValues(operation).Inc()
+		tuplesCacheHitCounter.WithLabelValues(operation, c.method).Inc()
 		span.SetAttributes(attribute.Bool("cached", true))
 
 		staticIter := storage.NewStaticIterator[*storage.TupleRecord](cacheEntry.Tuples)
@@ -401,6 +381,7 @@ func (c *CachedDatastore) newCachedIterator(
 		iter:      iter,
 		store:     store,
 		operation: operation,
+		method:    c.method,
 		// set an initial fraction capacity to balance constant reallocation and memory usage
 		tuples:            make([]*openfgav1.Tuple, 0, c.maxResultSize/2),
 		cacheKey:          cacheKey,
@@ -423,6 +404,7 @@ type cachedIterator struct {
 	iter              storage.TupleIterator
 	store             string
 	operation         string
+	method            string
 	cacheKey          string
 	invalidEntityKeys []string
 	cache             storage.InMemoryCache[any]
@@ -483,6 +465,7 @@ func (c *cachedIterator) Next(ctx context.Context) (*openfgav1.Tuple, error) {
 	if c.tuples != nil {
 		c.tuples = append(c.tuples, t)
 		if len(c.tuples) >= c.maxResultSize {
+			tuplesCacheDiscardCounter.WithLabelValues(c.operation, c.method).Inc()
 			c.tuples = nil // don't store results that are incomplete
 		}
 	}
@@ -614,7 +597,7 @@ func (c *cachedIterator) addToBuffer(t *openfgav1.Tuple) bool {
 	c.records = append(c.records, record)
 
 	if len(c.records) >= c.maxResultSize {
-		tuplesCacheDiscardCounter.WithLabelValues(c.operation).Inc()
+		tuplesCacheDiscardCounter.WithLabelValues(c.operation, c.method).Inc()
 		c.tuples = nil
 		c.records = nil
 	}
@@ -644,5 +627,5 @@ func (c *cachedIterator) flush() {
 	for _, k := range c.invalidEntityKeys {
 		c.cache.Delete(k)
 	}
-	tuplesCacheSizeHistogram.WithLabelValues(c.operation).Observe(float64(len(records)))
+	tuplesCacheSizeHistogram.WithLabelValues(c.operation, c.method).Observe(float64(len(records)))
 }